TMDM-6169: * DSC: New utility class to serialize or deserialize a task to/from XML * DSC: Updates to the XSD of a task (missing elements in XML format). * DSC: Adds new webservice that allows retrieval of the full task as XML. * DSC: Fix incorrect task update code from previous change. * DSC: Minor code improvements (e.g. usage of method overloading when appropriate). * MDM: Fix incorrect number comparison during golden record creation. * MDM: Restore "edit on staging resets status" (also improves design and separation of operations). * MDM: Fix incorrect query handling if query contains a taskId. * MDM: Adds resolution code (create a master record from user specified values in DSC). * MDM: Adds dependency to datastewardship 'client'.
git-svn-id: http://talendforge.org/svn/tom/trunk@107285 f6f1c999-d317-4740-80b0-e6d1abc6f99e

+++ org.talend.mdm.core/src/com/amalto/core/history/action/NumberActions.java
@@ -19,8 +19,8 @@ public class NumberActions {
     public static Action max(FieldUpdateAction action) {
         String newValueAsString = action.getNewValue();
         String oldValueAsString = action.getOldValue();
-        long newValue = newValueAsString.isEmpty() ? Long.MIN_VALUE : Long.parseLong(newValueAsString);
-        long oldValue = oldValueAsString.isEmpty() ? Long.MIN_VALUE : Long.parseLong(oldValueAsString);
+        float newValue = newValueAsString.isEmpty() ? Long.MIN_VALUE : Float.parseFloat(newValueAsString);
+        float oldValue = oldValueAsString.isEmpty() ? Long.MIN_VALUE : Float.parseFloat(oldValueAsString);
         if (newValue > oldValue) {
             return action;
         } else {
@@ -31,8 +31,8 @@ public class NumberActions {
     public static Action min(FieldUpdateAction action) {
         String newValueAsString = action.getNewValue();
         String oldValueAsString = action.getOldValue();
-        long newValue = newValueAsString.isEmpty() ? Long.MAX_VALUE : Long.parseLong(newValueAsString);
-        long oldValue = oldValueAsString.isEmpty() ? Long.MAX_VALUE : Long.parseLong(oldValueAsString);
+        float newValue = newValueAsString.isEmpty() ? Long.MAX_VALUE : Float.parseFloat(newValueAsString);
+        float oldValue = oldValueAsString.isEmpty() ? Long.MAX_VALUE : Float.parseFloat(oldValueAsString);
         if (newValue < oldValue) {
             return action;
         } else {
